+++ examples/min-dom/vite.config.ts
@@ -1,21 +1,8 @@
 import { defineConfig } from 'vite';
 
-// import babel from 'vite-plugin-babel';
-
 export default defineConfig({
-  // plugins: [
-  //   babel({
-  //     babelConfig: {
-  //       plugins: [
-  //         [
-  //           '@babel/plugin-transform-react-jsx',
-  //           {
-  //             pragma: 'h', // 自定义 JSX 编译函数
-  //             pragmaFrag: 'Fragment', // 自定义 Fragment
-  //           },
-  //         ],
-  //       ],
-  //     },
-  //   }),
-  // ],
+  esbuild: {
+    jsxImportSource: '@simple-data-open/min-dom',
+    jsx: 'automatic',
+  },
 });