+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @simple-data-open/types
 
+## 0.0.13
+
+### Patch Changes
+
+- 优化更新
+
 ## 0.0.12
 
 ### Patch Changes