+++ tools/util-pack/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @simple-data-open/util-pack
 
+## 0.0.7
+
+### Patch Changes
+
+- 优化更新
+
 ## 0.0.6
 
 ### Patch Changes